+++ mock/testcase/server/middleware/adapter.go
